grade range percent criteria inherits from graede range criteria
